test: migrate history-query itests to fullnode-specific (#1126)
